Change certain int fields to uint for validation

+++ garden-app/pkg/garden.go
@@ -23,7 +23,7 @@ type Garden struct {
 	Name          string            `json:"name" yaml:"name,omitempty"`
 	ID            xid.ID            `json:"id" yaml:"id,omitempty"`
 	Plants        map[xid.ID]*Plant `json:"plants" yaml:"plants,omitempty"`
-	MaxPlants     *int              `json:"max_plants" yaml:"max_plants"`
+	MaxPlants     *uint             `json:"max_plants" yaml:"max_plants"`
 	CreatedAt     *time.Time        `json:"created_at" yaml:"created_at,omitempty"`
 	EndDate       *time.Time        `json:"end_date,omitempty" yaml:"end_date,omitempty"`
 	LightSchedule *LightSchedule    `json:"light_schedule,omitempty" yaml:"light_schedule,omitempty"`
@@ -114,8 +114,8 @@ func (g *Garden) Patch(newGarden *Garden) {
 }
 
 // NumPlants returns the number of non-end-dated Plants that are part of this Garden
-func (g *Garden) NumPlants() int {
-	result := 0
+func (g *Garden) NumPlants() uint {
+	result := uint(0)
 	for _, p := range g.Plants {
 		if !p.EndDated() {
 			result++
