+++ garden-app/pkg/plant_test.go
@@ -42,7 +42,7 @@ func TestPlantEndDated(t *testing.T) {
 }
 
 func TestPlantPatch(t *testing.T) {
-	zero := 0
+	zero := uint(0)
 	now := time.Now()
 	tests := []struct {
 		name     string
